+++ module/CLI/src/Command/ShortUrl/CreateShortUrlCommand.php
@@ -4,24 +4,18 @@ declare(strict_types=1);
 
 namespace Shlinkio\Shlink\CLI\Command\ShortUrl;
 
+use Shlinkio\Shlink\CLI\Input\ShortUrlDataInput;
 use Shlinkio\Shlink\CLI\Util\ExitCode;
 use Shlinkio\Shlink\Core\Exception\NonUniqueSlugException;
 use Shlinkio\Shlink\Core\Options\UrlShortenerOptions;
 use Shlinkio\Shlink\Core\ShortUrl\Helper\ShortUrlStringifierInterface;
-use Shlinkio\Shlink\Core\ShortUrl\Model\ShortUrlCreation;
-use Shlinkio\Shlink\Core\ShortUrl\Model\Validation\ShortUrlInputFilter;
 use Shlinkio\Shlink\Core\ShortUrl\UrlShortenerInterface;
 use Symfony\Component\Console\Command\Command;
-use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
 
-use function array_map;
-use function array_unique;
-use function explode;
-use function Shlinkio\Shlink\Core\ArrayUtils\flatten;
 use function sprintf;
 
 class CreateShortUrlCommand extends Command
@@ -29,6 +23,7 @@ class CreateShortUrlCommand extends Command
     public const NAME = 'short-url:create';
 
     private ?SymfonyStyle $io;
+    private readonly ShortUrlDataInput $shortUrlDataInput;
 
     public function __construct(
         private readonly UrlShortenerInterface $urlShortener,
@@ -36,6 +31,7 @@ class CreateShortUrlCommand extends Command
         private readonly UrlShortenerOptions $options,
     ) {
         parent::__construct();
+        $this->shortUrlDataInput = new ShortUrlDataInput($this);
     }
 
     protected function configure(): void
@@ -43,26 +39,11 @@ class CreateShortUrlCommand extends Command
         $this
             ->setName(self::NAME)
             ->setDescription('Generates a short URL for provided long URL and returns it')
-            ->addArgument('longUrl', InputArgument::REQUIRED, 'The long URL to parse')
             ->addOption(
-                'tags',
-                't',
-                InputOption::VALUE_IS_ARRAY | InputOption::VALUE_REQUIRED,
-                'Tags to apply to the new short URL',
-            )
-            ->addOption(
-                'valid-since',
-                's',
+                'domain',
+                'd',
                 InputOption::VALUE_REQUIRED,
-                'The date from which this short URL will be valid. '
-                . 'If someone tries to access it before this date, it will not be found.',
-            )
-            ->addOption(
-                'valid-until',
-                'u',
-                InputOption::VALUE_REQUIRED,
-                'The date until which this short URL will be valid. '
-                . 'If someone tries to access it after this date, it will not be found.',
+                'The domain to which this short URL will be attached.',
             )
             ->addOption(
                 'custom-slug',
@@ -71,46 +52,22 @@ class CreateShortUrlCommand extends Command
                 'If provided, this slug will be used instead of generating a short code',
             )
             ->addOption(
-                'path-prefix',
-                'p',
-                InputOption::VALUE_REQUIRED,
-                'Prefix to prepend before the generated short code or provided custom slug',
-            )
-            ->addOption(
-                'max-visits',
-                'm',
-                InputOption::VALUE_REQUIRED,
-                'This will limit the number of visits for this short URL.',
-            )
-            ->addOption(
-                'find-if-exists',
-                'f',
-                InputOption::VALUE_NONE,
-                'This will force existing matching URL to be returned if found, instead of creating a new one.',
-            )
-            ->addOption(
-                'domain',
-                'd',
-                InputOption::VALUE_REQUIRED,
-                'The domain to which this short URL will be attached.',
-            )
-            ->addOption(
                 'short-code-length',
                 'l',
                 InputOption::VALUE_REQUIRED,
                 'The length for generated short code (it will be ignored if --custom-slug was provided).',
             )
             ->addOption(
-                'crawlable',
-                'r',
-                InputOption::VALUE_NONE,
-                'Tells if this URL will be included as "Allow" in Shlink\'s robots.txt.',
+                'path-prefix',
+                'p',
+                InputOption::VALUE_REQUIRED,
+                'Prefix to prepend before the generated short code or provided custom slug',
             )
             ->addOption(
-                'no-forward-query',
-                'w',
+                'find-if-exists',
+                'f',
                 InputOption::VALUE_NONE,
-                'Disables the forwarding of the query string to the long URL, when the new short URL is visited.',
+                'This will force existing matching URL to be returned if found, instead of creating a new one.',
             );
     }
 
@@ -136,32 +93,17 @@ class CreateShortUrlCommand extends Command
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
         $io = $this->getIO($input, $output);
-        $longUrl = $input->getArgument('longUrl');
-        if (empty($longUrl)) {
-            $io->error('A URL was not provided!');
-            return ExitCode::EXIT_FAILURE;
-        }
-
-        $explodeWithComma = static fn (string $tag) => explode(',', $tag);
-        $tags = array_unique(flatten(array_map($explodeWithComma, $input->getOption('tags'))));
-        $maxVisits = $input->getOption('max-visits');
-        $shortCodeLength = $input->getOption('short-code-length') ?? $this->options->defaultShortCodesLength;
 
         try {
-            $result = $this->urlShortener->shorten(ShortUrlCreation::fromRawData([
-                ShortUrlInputFilter::LONG_URL => $longUrl,
-                ShortUrlInputFilter::VALID_SINCE => $input->getOption('valid-since'),
-                ShortUrlInputFilter::VALID_UNTIL => $input->getOption('valid-until'),
-                ShortUrlInputFilter::MAX_VISITS => $maxVisits !== null ? (int) $maxVisits : null,
-                ShortUrlInputFilter::CUSTOM_SLUG => $input->getOption('custom-slug'),
-                ShortUrlInputFilter::PATH_PREFIX => $input->getOption('path-prefix'),
-                ShortUrlInputFilter::FIND_IF_EXISTS => $input->getOption('find-if-exists'),
-                ShortUrlInputFilter::DOMAIN => $input->getOption('domain'),
-                ShortUrlInputFilter::SHORT_CODE_LENGTH => $shortCodeLength,
-                ShortUrlInputFilter::TAGS => $tags,
-                ShortUrlInputFilter::CRAWLABLE => $input->getOption('crawlable'),
-                ShortUrlInputFilter::FORWARD_QUERY => !$input->getOption('no-forward-query'),
-            ], $this->options));
+            $result = $this->urlShortener->shorten($this->shortUrlDataInput->toShortUrlCreation(
+                $input,
+                $this->options,
+                customSlugField: 'custom-slug',
+                shortCodeLengthField: 'short-code-length',
+                pathPrefixField: 'path-prefix',
+                findIfExistsField: 'find-if-exists',
+                domainField: 'domain',
+            ));
 
             $result->onEventDispatchingError(static fn () => $io->isVerbose() && $io->warning(
                 'Short URL properly created, but the real-time updates cannot be notified when generating the '
@@ -169,7 +111,7 @@ class CreateShortUrlCommand extends Command
             ));
 
             $io->writeln([
-                sprintf('Processed long URL: <info>%s</info>', $longUrl),
+                sprintf('Processed long URL: <info>%s</info>', $result->shortUrl->getLongUrl()),
                 sprintf('Generated short URL: <info>%s</info>', $this->stringifier->stringify($result->shortUrl)),
             ]);
             return ExitCode::EXIT_SUCCESS;
@@ -181,6 +123,6 @@ class CreateShortUrlCommand extends Command
 
     private function getIO(InputInterface $input, OutputInterface $output): SymfonyStyle
     {
-        return $this->io ?? ($this->io = new SymfonyStyle($input, $output));
+        return $this->io ??= new SymfonyStyle($input, $output);
     }
 }