+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 * [#2018](https://github.com/shlinkio/shlink/issues/2018) Add option to allow all short URLs to be unconditionally crawlable in robots.txt, via `ROBOTS_ALLOW_ALL_SHORT_URLS=true` env var, or config options.
 
 ### Changed
-* *Nothing*
+* [#2096](https://github.com/shlinkio/shlink/issues/2096) Update to RoadRunner 2024.
 
 ### Deprecated
 * *Nothing*